Minor counter tweaks for test coverage and minor signature changes

+++ hazelcast/src/main/java/com/hazelcast/util/counters/Counter.java
@@ -45,5 +45,5 @@ public interface Counter {
      *
      * @param amount the amount to increase or decrease the counter with.
      */
-    void inc(int amount);
+    void inc(long amount);
 }
